requirements.txt
fastapi
uvicorn[standard]
pydantic
joblib
numpy
pandas
yfinance
shap
matplotlib
pygooglenews
newspaper3k
nltk